refactor(indexeddb): move SafeEncode-related traits and types into their own module
