+++ crates/matrix-sdk-indexeddb/src/lib.rs
@@ -10,7 +10,6 @@ mod crypto_store;
 mod event_cache_store;
 #[cfg(feature = "media-store")]
 mod media_store;
-mod safe_encode;
 #[cfg(feature = "e2e-encryption")]
 mod serialize_bool_for_indexeddb;
 #[cfg(feature = "e2e-encryption")]
