+++ crates/matrix-sdk-indexeddb/src/serializer/safe_encode/traits.rs
